Extract method for detect social image in page.
